disabled top and bottom limits in the Elevator. We think the encoders are preventing the elevator from going down but not sure why.
Added a band to the elevator speed so if the speed calculated from the toggles are less that .1 and greater than -0.1 then set hold flag and tell the elevator subsystem to hold at the current position.

Before this change any negative speed to the elevator activates the hold system and the elevator won't go down.